refactor: update parent detection
Previously we do searched all potential open pull requests with a
managed label and then check their references. This has a lot of
drawbacks (e.g. if the managed pr is closed).

The new approach is to iterate all prs and select the first one which
has a reference to the review PR in question. This does also take
closed PRs into account and seems to be way more stable.

This may still lead to wrong detection of parent PR in some cases
(e.g. if a review PR is referenced in an unrelated PR with the same
base).