feat: allow tasks to be canceled
There are cases where tasks should superseded by other tasks being
scheduled (examples):

- A sync is running, because a review was merged. While this is
  executing another review is merged -> the current job can
  be canceled and superseded by a new sync task.
- A managed pr was closed -> all current updates can be deferred to
  the cleanup task.

Unfortunately it is not easily doable to use generators and types in
combination (we do want to stay typesafe here), we did implement a
cancellation token pattern.
Before each asynchronous operation a check is done if the current task
is canceled. In that case the whole job aborted.