feat: update pull requests on scheduled base
In case of missing webhooks, we need to synchronize the pull requests
on a regular base. This is implemented in this pull request.

It could be managed by setting SCHEDULE_DELAY and SCHEDULE_TIMEOUT.
The delay is an initial wait interval, the timeout is the time between
scheduled update runs.